+++ websub-compiler-plugin-test/src/test/resources/ballerina_sources/sample_16/Ballerina.toml
@@ -1,0 +1,7 @@
+[package]
+org = "websub_test"
+name = "sample_16"
+version = "0.1.0"
+
+[build-options]
+observabilityIncluded = true
